feature/orderService/ cancelOrder SHIPPED 상태이면 수정 불가

+++ backend/project1/src/main/java/com/team4/project1/domain/order/service/OrderService.java
@@ -101,6 +101,13 @@ public class OrderService {
 
     // 주문 취소 메소드
     public Long cancelOrder(Long orderId) {
+        Order existingOrder = orderRepository.findById(orderId)
+                        .orElseThrow(() -> new RuntimeException("주문을 찾을 수 없습니다. (ID: " + orderId + ")"));
+
+        if(existingOrder.getDeliveryStatus() == DeliveryStatus.SHIPPED) {
+            throw new IllegalStateException(("이미 발송된 주문은 취소할 수 없습니다."));
+        }
+
         orderRepository.deleteById(orderId);
         return orderId;
     }
